requirements.txt
# Dependências do Python para o projeto
mysql-connector-python
python-dateutil